HighScore Scene
Ahora muestra el record

+++ PR_ZAXXON_HolgueraAlvaro/Assets/Scripts/Config/GameManager.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 public class GameManager : MonoBehaviour
 {
     public static float volumeMusic = 5f;
-    public static int highScore = 0;
+    public static float highScore = 0;
 }
